// frontend/src/groups/vcsprovider/__generated__/VCSProviderListFragment_vcsProviders.graphql.ts
export type VCSProviderListFragment_vcsProviders$data = {
  readonly id: string;
  readonly vcsProviders: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly groupPath: string;
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"VCSProviderListItemFragment_vcsProvider">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "VCSProviderListFragment_vcsProviders";
};
export type VCSProviderListFragment_vcsProviders$key = {
  readonly " $data"?: VCSProviderListFragment_vcsProviders$data;
  readonly " $fragmentSpreads": FragmentRefs<"VCSProviderListFragment_vcsProviders">;
};